sanitize part 2
git-svn-id: https://plugins.svn.wordpress.org/daily-menu/trunk@1073365 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ models/Dish.php
@@ -110,6 +110,7 @@ class Dish {
 	 */
 	function removeFromMenu($date) {
 		global $wpdb;
+		if (!isset($date)) return false;
 		if (!isset($this->id_dish)) return false;
 		return $wpdb->delete($wpdb->prefix . Menu::getTableSuffix(), array( 'DATE' => $date , 'ID_DISH' => $this->id_dish));	
 	}
